Se hicieron cambios para el inicio de administrador y dependiendo su rol le mostrara ó ocultará algunos botones de la interfaz.
NOTA!!!: Aun falta un cambio porque se tiene un herror de css en un boton de la interfaz de usuario
