+++ bases/mainInterfaz/footer.php
@@ -2,5 +2,6 @@
     </div> 
     <script src="../../../Proyecto_SendApp_2024/scripts/menuHamburguesaJS/menuHaburguesaInterface.js"></script>
     <script src="../../../Proyecto_SendApp_2024/bases/mainInterfaz/actualizarImagen/actualizar.js"></script>
+    <script src="../../../Proyecto_SendApp_2024/interfaces/Administrador/Scripts/admin.js"></script>
 </body>
 </html>
